+++ src/lib/CommandBase.ts
@@ -11,9 +11,10 @@ export abstract class CommandBase {
 		const configuration = workspace.getConfiguration();
 		const defaultStage: string = configuration.get("serverless.aws.defaultStage") || "dev";
 		const defaultRegion: string = configuration.get("serverless.aws.defaultRegion") || "ap-northeast-2";
-		const defaultNodeModulesPath: string = configuration.get("serverless.defaultNodeModulesPath") || '/usr/local/lib/node_modules';
+		const defaultNodeModulesPath: string = configuration.get("serverless.defaultNodeModulesPath") || "/usr/local/lib/node_modules";
+		const credentialName: string = configuration.get( `serverless.aws.credentials.${defaultStage}`) || "default";
 
-		return Promise.resolve([ defaultStage, defaultRegion, defaultNodeModulesPath ]);
+		return Promise.resolve( [ defaultStage, defaultRegion, defaultNodeModulesPath, credentialName ]);
 	}
 
 	constructor(public readonly isExclusive: boolean = false) {