+++ src/lib/commands/Logs.ts
@@ -23,12 +23,14 @@ export class Logs extends CommandBase {
 		return CommandBase.askForStageAndRegion()
 		.then(result => {
 			const options = {
-				cwd: node.documentRoot,
-				function: node.name,
-				region: result[1],
-				stage: result[ 0 ]
+				'cwd': node.documentRoot,
+				'function': node.name,
+				'region': result[1],
+				'stage': result[0],
+				'aws-profile': result[3]
 			};
-			return Serverless.invoke( "logs", options, result[ 2 ]);
+
+			return Serverless.invoke( "logs", options, result[2]);
 		});
 	}