+++ src/lib/commands/Package.ts
@@ -23,10 +23,12 @@ export class Package extends CommandBase {
 		return CommandBase.askForStageAndRegion()
 		.then(result => {
 			const options = {
-				cwd: node.documentRoot,
-				region: result[1],
-				stage: result[ 0 ]
+				'cwd': node.documentRoot,
+				'region': result[1],
+				'stage': result[0],
+				'aws-profile': result[3]
 			};
+
 			return Serverless.invoke( "package", options, result[ 2 ]);
 		});
 	}