+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.0.6
+- AWS Credentials (prod, dev) 로 다중설정할 수 있도록 설정 추가
+
 ## 1.0.5
 - Invoke function 옵션 추가
 - Deploy service 하지 않은 상태에서 Invoke function 호출시 반응이 없을때 프로세스 중지 안내 추가